make a separate DropdownMultiple component
