+++ spotlight-client/src/UiLibrary/Dropdown/Dropdown.test.tsx
@@ -16,13 +16,11 @@
 // =============================================================================
 
 import { render, fireEvent, screen, waitFor } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
 import React from "react";
 import Dropdown from "./Dropdown";
 
 const testLabel = "Test Label";
 const mockOnChange = jest.fn();
-const mockOnHighlight = jest.fn();
 const testOptions = [
   { id: "1", label: "option one" },
   { id: "2", label: "option two" },
@@ -150,121 +148,4 @@ test("options can be hidden", () => {
   );
 
   expect(menuButton).toHaveTextContent(hiddenOption.label);
-});
-
-describe("multiple select", () => {
-  let rendered: ReturnType<typeof render>;
-
-  beforeEach(() => {
-    rendered = render(
-      <Dropdown
-        label={testLabel}
-        multiple
-        options={testOptions}
-        onChangeMultiple={mockOnChange}
-        onHighlight={mockOnHighlight}
-        selectedId={["1", "2", "3"]}
-      />
-    );
-  });
-
-  test("selects multiple", () => {
-    const menuButton = screen.getByRole("button", {
-      name: `${testLabel} ${testOptions[0].label} and 2 others`,
-    });
-
-    fireEvent.click(menuButton);
-
-    testOptions.forEach((opt) => {
-      expect(
-        screen.getByRole("option", { name: opt.label, selected: true })
-      ).toBeVisible();
-    });
-  });
-
-  test("menu stays open after selection", async () => {
-    const menuButton = screen.getByRole("button", {
-      name: `${testLabel} ${testOptions[0].label} and 2 others`,
-    });
-
-    fireEvent.click(menuButton);
-
-    fireEvent.click(screen.getByRole("option", { name: testOptions[0].label }));
-    expect(mockOnChange).toHaveBeenLastCalledWith([
-      testOptions[1].id,
-      testOptions[2].id,
-    ]);
-    // update the controlled value to simulate real use
-    rendered.rerender(
-      <Dropdown
-        label={testLabel}
-        multiple
-        options={testOptions}
-        onChangeMultiple={mockOnChange}
-        selectedId={["2", "3"]}
-      />
-    );
-
-    // let's give the animation some time to run;
-    // its duration is not entirely predictable but this should be enough time
-    await new Promise((resolve) => setTimeout(resolve, 750));
-
-    const option2 = screen.getByRole("option", { name: testOptions[1].label });
-    expect(option2).toBeVisible();
-
-    fireEvent.click(option2);
-    expect(mockOnChange).toHaveBeenLastCalledWith([testOptions[2].id]);
-  });
-
-  test("select all", () => {
-    const menuButton = screen.getByRole("button", {
-      name: `${testLabel} ${testOptions[0].label} and 2 others`,
-    });
-
-    fireEvent.click(menuButton);
-
-    const selectAll = screen.getByRole("option", { name: "Deselect all" });
-    expect(selectAll).toBeVisible();
-
-    fireEvent.click(selectAll);
-    expect(mockOnChange).toHaveBeenLastCalledWith([]);
-
-    // update the controlled value to simulate real use; none are selected
-    rendered.rerender(
-      <Dropdown
-        label={testLabel}
-        multiple
-        options={testOptions}
-        onChangeMultiple={mockOnChange}
-        selectedId={[]}
-      />
-    );
-
-    expect(selectAll).toHaveTextContent("Select all");
-    expect(
-      screen.queryByRole("option", { selected: true })
-    ).not.toBeInTheDocument();
-
-    fireEvent.click(selectAll);
-    expect(mockOnChange).toHaveBeenLastCalledWith(
-      testOptions.map(({ id }) => id)
-    );
-  });
-
-  test("callback for highlighted item", async () => {
-    const menuButton = screen.getByRole("button", {
-      name: `${testLabel} ${testOptions[0].label} and 2 others`,
-    });
-
-    userEvent.click(menuButton);
-
-    userEvent.hover(
-      await screen.findByRole("option", { name: "Deselect all" })
-    );
-    expect(mockOnHighlight).not.toHaveBeenCalled();
-
-    userEvent.hover(screen.getByRole("option", { name: testOptions[0].label }));
-
-    expect(mockOnHighlight).toHaveBeenLastCalledWith(testOptions[0].id);
-  });
 });
